src/requirements.txt
inject
boto3
PyJWT
requests
bcrypt
botocore